(chore): moved typings to decleration file
